+++ kernel/include/kernel_internal.h
@@ -110,15 +110,6 @@ static inline void *z_thread_malloc(size_t size)
 	return z_thread_aligned_alloc(0, size);
 }
 
-/* clean up when a thread is aborted */
-
-#if defined(CONFIG_THREAD_MONITOR)
-extern void z_thread_monitor_exit(struct k_thread *thread);
-#else
-#define z_thread_monitor_exit(thread) \
-	do {/* nothing */    \
-	} while (false)
-#endif /* CONFIG_THREAD_MONITOR */
 
 #ifdef CONFIG_USE_SWITCH
 /* This is a arch function traditionally, but when the switch-based
